CHeli.h: Fixed types of m_fRotorZ and m_fSecondRotorZ

+++ plugin_sa/game_sa/CHeli.h
@@ -60,8 +60,8 @@ public:
     float              m_fSteeringLeftRight;
     float              m_fAccelerationBreakStatus;
     int field_99C;
-    int                m_fRotorZ;
-    int                m_fSecondRotorZ;
+    float              m_fRotorZ;
+    float              m_fSecondRotorZ;
     float              m_fMaxAltitude;
     float field_9AC;
     float              m_fMinAltitude;
